test(sessions): align remaining harnesses with read-only accessors
